refactor(dispatcher): extract CQRS state projector and restore domain context (#979)
