+++ src/ramses_rf/state_projector.py
@@ -1,0 +1,749 @@
+#!/usr/bin/env python3
+"""RAMSES RF - CQRS State Projector for mapping telemetry to read-models."""
+
+from __future__ import annotations
+
+import contextlib
+import dataclasses
+import logging
+import uuid
+from collections.abc import Callable
+from typing import TYPE_CHECKING, Any, Final
+
+from . import exceptions as exc, quirks
+from .const import (
+    SZ_ACTIVE,
+    SZ_AIR_QUALITY,
+    SZ_AIR_QUALITY_BASIS,
+    SZ_BYPASS_MODE,
+    SZ_BYPASS_POSITION,
+    SZ_BYPASS_STATE,
+    SZ_CO2_LEVEL,
+    SZ_DATETIME,
+    SZ_DIFFERENTIAL,
+    SZ_EXHAUST_FAN_SPEED,
+    SZ_EXHAUST_FLOW,
+    SZ_EXHAUST_TEMP,
+    SZ_FAN_INFO,
+    SZ_FAN_MODE,
+    SZ_FAN_RATE,
+    SZ_FILTER_DIRTY,
+    SZ_FROST_CYCLE,
+    SZ_HAS_FAULT,
+    SZ_HEAT_DEMAND,
+    SZ_INDOOR_HUMIDITY,
+    SZ_INDOOR_TEMP,
+    SZ_LANGUAGE,
+    SZ_MINUTES,
+    SZ_MODE,
+    SZ_OUTDOOR_HUMIDITY,
+    SZ_OUTDOOR_TEMP,
+    SZ_OVERRUN,
+    SZ_POST_HEAT,
+    SZ_PRE_HEAT,
+    SZ_PRESENCE_DETECTED,
+    SZ_RELAY_DEMAND,
+    SZ_REMAINING_DAYS,
+    SZ_REMAINING_MINS,
+    SZ_REMAINING_PERCENT,
+    SZ_REQ_REASON,
+    SZ_REQ_SPEED,
+    SZ_SETPOINT,
+    SZ_SPEED_CAPABILITIES,
+    SZ_SUPPLY_FAN_SPEED,
+    SZ_SUPPLY_FLOW,
+    SZ_SUPPLY_TEMP,
+    SZ_SYSTEM_MODE,
+    SZ_TEMPERATURE,
+    SZ_UNTIL,
+    Code,
+)
+from .devices.hvac_ventilators import HvacVentilator
+from .messages import Message
+from .models import StateUpdatedEvent, SystemState
+from .systems.faultlog import FaultLogEntry
+from .systems.zones import DhwZone
+from .topology_builder import update_topology_schema_state
+
+if TYPE_CHECKING:
+    from .gateway import Gateway
+
+_LOGGER = logging.getLogger(__name__)
+
+__all__ = [
+    "StateProjector",
+    "StateProjectorRegistry",
+    "process_state_updates",
+]
+
+
+class StateProjectorRegistry:
+    """Registry mapping Code opcode enums to specific state updater functions."""
+
+    def __init__(self) -> None:
+        """Initialize the state projector registry."""
+        self._handlers: dict[
+            Code, list[Callable[[Any, dict[str, Any], Message], None]]
+        ] = {}
+
+    def register(
+        self,
+        code: Code,
+        handler: Callable[[Any, dict[str, Any], Message], None],
+    ) -> None:
+        """Register an updater handler for a specific opcode.
+
+        :param code: The packet opcode to handle.
+        :type code: Code
+        :param handler: The updater function to invoke.
+        :type handler: Callable[[Any, dict[str, Any], Message], None]
+        """
+        self._handlers.setdefault(code, []).append(handler)
+
+    def get_handlers(
+        self, code: Code
+    ) -> list[Callable[[Any, dict[str, Any], Message], None]]:
+        """Retrieve registered handlers for an opcode.
+
+        :param code: The packet opcode to look up.
+        :type code: Code
+        :return: List of updater callables for the opcode.
+        :rtype: list[Callable[[Any, dict[str, Any], Message], None]]
+        """
+        return self._handlers.get(code, [])
+
+
+class StateProjector:
+    """CQRS State Projector for translating raw payloads into read-models."""
+
+    def __init__(self, gwy: Gateway) -> None:
+        """Initialize the state projector with a gateway instance.
+
+        :param gwy: The gateway handling device entities.
+        :type gwy: Gateway
+        """
+        self._gwy = gwy
+        self._registry = StateProjectorRegistry()
+        self._setup_registry()
+
+    def _setup_registry(self) -> None:
+        """Populate opcode handlers into the registry."""
+        self._registry.register(Code._0100, _update_system_state)
+        self._registry.register(Code._2E04, _update_system_state)
+        self._registry.register(Code._313F, _update_system_state)
+
+        self._registry.register(Code._10A0, _update_dhw_state)
+        self._registry.register(Code._1F41, _update_dhw_state)
+
+        self._registry.register(Code._0006, _update_schedule_state)
+        self._registry.register(Code._0404, _update_schedule_state)
+
+        self._registry.register(Code._0418, _update_faultlog_state)
+
+    async def process_msg(self, msg: Message) -> None:
+        """Project a decoded message into read-models.
+
+        :param msg: The message containing payload telemetry.
+        :type msg: Message
+        """
+        await process_state_updates(self._gwy, msg)
+
+
+_DHW_OPCODES: Final[frozenset[Code]] = frozenset({Code._1260, Code._10A0, Code._1F41})
+
+
+def _get_dhw_zone_from_msg(msg: Message, src_dev: Any) -> DhwZone | None:
+    """Resolve the DhwZone that should ingest a DHW opcode (1260/10A0/1F41).
+
+    These payloads carry no ``zone_idx``/``domain_id``, so standard target
+    resolution in ``_resolve_logical_targets`` misses the DhwZone.
+
+    ``1260`` is sent by the DhwSensor (or relayed by the Controller as an
+    RP); ``10A0``/``1F41`` are sent by the Controller.  The
+    appliance_control (OTB) also emits ``10A0``/``1260`` with different
+    semantics (CH setpoint / null temp) and is excluded to avoid
+    clobbering the DHW read-models.
+
+    See: https://github.com/ramses-rf/ramses_cc/issues/843
+
+    :param msg: The inbound message.
+    :type msg: Message
+    :param src_dev: The source device (DhwSensor or Controller).
+    :type src_dev: Any
+    :return: The DhwZone to route to, or ``None`` if the message is not
+        a DHW opcode or the source is not a DHW sender.
+    :rtype: DhwZone | None
+    """
+    if msg.code not in _DHW_OPCODES or src_dev is None:
+        return None
+
+    src_slug = getattr(src_dev, "_SLUG", "")
+    if msg.code == Code._1260:
+        is_dhw_src = src_slug in ("DHW", "CTL")
+    else:  # 10A0 / 1F41 are owned by the Controller
+        is_dhw_src = src_slug == "CTL"
+
+    if not is_dhw_src:
+        return None
+
+    tcs = getattr(src_dev, "tcs", None)
+    if tcs is None:
+        return None
+
+    return getattr(tcs, "dhw", None)
+
+
+def _resolve_logical_targets(
+    gwy: Gateway, msg: Message, p: dict[str, Any]
+) -> list[Any]:
+    """Resolve software twin entities targeted by a payload.
+
+    :param gwy: Gateway instance with device registry.
+    :type gwy: Gateway
+    :param msg: L7 Message envelope.
+    :type msg: Message
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :return: List of target entity instances.
+    :rtype: list[Any]
+    """
+    targets: list[Any] = []
+    registry = getattr(gwy, "device_registry", None)
+    src_dev = registry.device_by_id.get(msg.src.id) if registry else None
+    dst_dev = (
+        registry.device_by_id.get(msg.dst.id)
+        if registry and hasattr(msg.dst, "id")
+        else None
+    )
+
+    tcs = getattr(src_dev, "tcs", None) if src_dev else None
+    tcs = tcs or getattr(gwy, "tcs", None)
+    if tcs is None and registry:
+        for dev in registry.device_by_id.values():
+            if str(dev.id).startswith("01:"):
+                tcs = getattr(dev, "tcs", None) or dev
+                break
+
+    # 1. Fault logs strictly target the TCS (if it exists) or the source device
+    if msg.code == Code._0418:
+        if tcs:
+            targets.append(getattr(tcs, "faultlog", src_dev))
+        elif src_dev:
+            targets.append(src_dev)
+        return targets
+
+    # 2. Hardware twin (Sender) always gets the update UNLESS it's a Controller/UFC
+    # actively broadcasting an array of children's states (e.g., a 30C9 sync).
+    src_type = getattr(src_dev, "type", None)
+    has_arr = getattr(msg, "_has_array", False)
+    if src_type not in ("01", "02") or not has_arr:
+        if src_dev:
+            targets.append(src_dev)
+
+    # 3. Hardware twin (Destination) gets the update.
+    # Legacy routes packets to the destination device's cache. To maintain
+    # strict parity, we mirror this.
+    # HVAC packets (e.g. 22F1 fan_mode from REM->FAN) target the destination
+    # device's hvac_state directly, so we also accept devices that have
+    # hvac_state even if they lack apply_state_update.
+    if msg.dst.id != msg.src.id and getattr(msg.dst, "id", "") != "63:262142":
+        if (
+            dst_dev
+            and (
+                getattr(dst_dev, "apply_state_update", None) is not None
+                or getattr(dst_dev, "hvac_state", None) is not None
+            )
+            and dst_dev not in targets
+        ):
+            targets.append(dst_dev)
+
+    # 4. Virtual twins (Zones) get updates if explicitly addressed by idx.
+    if "zone_idx" in p and tcs:
+        if zone := tcs.zone_by_idx.get(p["zone_idx"]):
+            if zone not in targets:
+                targets.append(zone)
+
+    # 5. Domain twins (TCS, DHW) get updates.
+    if "domain_id" in p and tcs:
+        domain_id = p["domain_id"]
+        if domain_id == "FC" and tcs not in targets:
+            targets.append(tcs)
+        elif domain_id in ("FA", "F9") and getattr(tcs, "dhw", None) is not None:
+            if tcs.dhw not in targets:
+                targets.append(tcs.dhw)
+
+    # 6. System-level opcodes (2E04/0100/313F) target the TCS directly.
+    #    These packets have no domain_id/zone_idx, so steps 4/5 miss them.
+    if msg.code in (Code._2E04, Code._0100, Code._313F) and tcs and tcs not in targets:
+        targets.append(tcs)
+
+    # 7. DHW opcodes (1260/10A0/1F41) carry no domain_id/zone_idx, so steps
+    #    4/5 miss the DhwZone.  Route them via the shared helper.
+    #    See: https://github.com/ramses-rf/ramses_cc/issues/843
+    dhw = _get_dhw_zone_from_msg(msg, src_dev)
+    if dhw is not None and dhw not in targets:
+        targets.append(dhw)
+
+    return targets
+
+
+def _update_system_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Translate system configuration opcodes into SystemState.
+
+    Handles 2E04 (system_mode), 0100 (language), and 313F (datetime).
+
+    :param target: Target entity (TCS/Evohome) to update.
+    :type target: Any
+    :param p: Parsed message payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Immutable Message envelope.
+    :type msg: Message
+    """
+    system_state = getattr(target, "system_state", None)
+    if system_state is None or not dataclasses.is_dataclass(system_state):
+        return
+
+    updates: dict[str, Any] = {}
+    if msg.code == Code._0100:
+        if SZ_LANGUAGE in p:
+            updates[SZ_LANGUAGE] = p[SZ_LANGUAGE]
+    elif msg.code == Code._2E04:
+        if SZ_SYSTEM_MODE in p:
+            updates[SZ_SYSTEM_MODE] = p[SZ_SYSTEM_MODE]
+        if SZ_UNTIL in p:
+            updates[SZ_UNTIL] = p[SZ_UNTIL]
+    elif msg.code == Code._313F:
+        if SZ_DATETIME in p:
+            updates[SZ_DATETIME] = p[SZ_DATETIME]
+    else:
+        return
+
+    if not updates:
+        return
+
+    dtm = getattr(msg, "dtm", getattr(msg, "timestamp", None))
+    if dtm:
+        updates["last_updated"] = dtm
+
+    current_state = target.system_state or SystemState()
+    new_state = dataclasses.replace(current_state, **updates)
+    target.system_state = new_state
+
+    event = StateUpdatedEvent(
+        entity_id=getattr(target, "id", "unknown"),
+        state=new_state,
+        correlation_id=getattr(msg, "correlation_id", uuid.uuid4()),
+        causation_id=getattr(msg, "message_id", uuid.uuid4()),
+    )
+    if hasattr(target, "apply_state_update"):
+        target.apply_state_update(event)
+
+
+def _update_hvac_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Translate HVAC ventilation payloads into a frozen HvacState.
+
+    Handles 31D9/31DA/22F1/22F3/10D0/12A0/1298 and related opcodes,
+    porting the logic into the CQRS state projector.
+    See issues ramses-rf/ramses_rf#649 and ramses-rf/ramses_rf#547.
+
+    :param target: Target entity to update.
+    :type target: Any
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    if getattr(target, "_SLUG", "") in ("CTL", "BDR", "TRV", "OTB", "UFC", "DHW"):
+        return
+
+    hvac_state = getattr(target, "hvac_state", None)
+    if hvac_state is None or not dataclasses.is_dataclass(hvac_state):
+        return
+
+    p = quirks.apply_hvac_quirks(p, target.hvac_state, msg.code)
+
+    fields = [
+        SZ_CO2_LEVEL,
+        SZ_AIR_QUALITY,
+        SZ_AIR_QUALITY_BASIS,
+        SZ_BYPASS_MODE,
+        SZ_BYPASS_POSITION,
+        SZ_BYPASS_STATE,
+        SZ_EXHAUST_FAN_SPEED,
+        SZ_EXHAUST_FLOW,
+        SZ_EXHAUST_TEMP,
+        SZ_FAN_RATE,
+        SZ_FAN_MODE,
+        SZ_FAN_INFO,
+        SZ_INDOOR_HUMIDITY,
+        SZ_INDOOR_TEMP,
+        SZ_OUTDOOR_HUMIDITY,
+        SZ_OUTDOOR_TEMP,
+        SZ_POST_HEAT,
+        SZ_PRE_HEAT,
+        SZ_PRESENCE_DETECTED,
+        SZ_REMAINING_MINS,
+        SZ_SPEED_CAPABILITIES,
+        SZ_SUPPLY_FAN_SPEED,
+        SZ_SUPPLY_FLOW,
+        SZ_SUPPLY_TEMP,
+        SZ_TEMPERATURE,
+        SZ_FILTER_DIRTY,
+        SZ_FROST_CYCLE,
+        SZ_HAS_FAULT,
+        "dewpoint_temp",
+    ]
+
+    _NULL_HUMIDITY_FIELDS = frozenset({SZ_INDOOR_HUMIDITY, SZ_OUTDOOR_HUMIDITY})
+
+    updates: dict[str, Any] = {}
+    for f in fields:
+        if f not in p:
+            continue
+        val = p[f]
+        # Filter out null-marker values that 31DA/31D9 snapshots emit for
+        # sensors the device does not have.  Without this, every polling cycle
+        # (~10 min) overwrites good telemetry from 22F1/12A0/22F7 with null
+        # markers, causing sensors to bounce to None/FF/0.  See issue #742.
+        if val is None:
+            continue
+        # None = "not implemented" (e.g. EF in bypass_position)
+        # Raw hex (e.g. "FF", "04") = non-semantic fan_mode from 31D9
+        # long-payload devices; the quirk normalises these to None, but
+        # filter here as belt-and-suspenders.  See ramses_cc issue 723.
+        if f == SZ_FAN_MODE and isinstance(val, str) and len(val) == 2:
+            try:
+                int(val, 16)
+                continue
+            except ValueError:
+                pass
+        # 0.0 for humidity = "no sensor" (00 parses as 0%, physically impossible)
+        if f in _NULL_HUMIDITY_FIELDS and val == 0:
+            continue
+        updates[f] = val
+
+    # Handle non-standard names passed by the semantic parsers
+    if SZ_REMAINING_DAYS in p and p[SZ_REMAINING_DAYS] is not None:
+        updates["filter_remaining_days"] = p[SZ_REMAINING_DAYS]
+    if SZ_REMAINING_PERCENT in p and p[SZ_REMAINING_PERCENT] is not None:
+        updates["filter_remaining_percent"] = p[SZ_REMAINING_PERCENT]
+    if SZ_MINUTES in p and msg.code == Code._22F3 and p[SZ_MINUTES] is not None:
+        updates["boost_timer_mins"] = p[SZ_MINUTES]
+    if SZ_REQ_SPEED in p and p[SZ_REQ_SPEED] is not None:
+        updates["request_fan_speed"] = p[SZ_REQ_SPEED]
+    if SZ_REQ_REASON in p and p[SZ_REQ_REASON] is not None:
+        updates["request_reason"] = p[SZ_REQ_REASON]
+
+    if not updates:
+        return
+
+    new_state = dataclasses.replace(target.hvac_state, **updates)
+    target.hvac_state = new_state
+
+    event = StateUpdatedEvent(
+        entity_id=getattr(target, "id", "unknown"),
+        state=new_state,
+        correlation_id=getattr(msg, "correlation_id", uuid.uuid4()),
+        causation_id=getattr(msg, "message_id", uuid.uuid4()),
+    )
+    if hasattr(target, "apply_state_update"):
+        target.apply_state_update(event)
+
+
+def _update_dhw_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Translate DHW opcodes (10A0/1260/1F41) into the frozen DhwState.
+
+    Hydrates the DhwZone's ``dhw_state`` read-model (setpoint/overrun/
+    differential from 10A0, mode/active/until from 1F41) in addition
+    to ``temp_state``.
+
+    :param target: Target entity to update.
+    :type target: Any
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    if not isinstance(target, DhwZone):
+        return
+    dhw_state = getattr(target, "dhw_state", None)
+    if dhw_state is None or not dataclasses.is_dataclass(dhw_state):
+        return
+
+    updates: dict[str, Any] = {}
+    if msg.code == Code._10A0:
+        if SZ_SETPOINT in p:
+            updates[SZ_SETPOINT] = p[SZ_SETPOINT]
+        if SZ_OVERRUN in p:
+            updates[SZ_OVERRUN] = p[SZ_OVERRUN]
+        if SZ_DIFFERENTIAL in p:
+            updates[SZ_DIFFERENTIAL] = p[SZ_DIFFERENTIAL]
+    elif msg.code == Code._1F41:
+        if SZ_MODE in p:
+            updates[SZ_MODE] = p[SZ_MODE]
+        if SZ_ACTIVE in p:
+            updates[SZ_ACTIVE] = p[SZ_ACTIVE]
+        if SZ_UNTIL in p:
+            updates[SZ_UNTIL] = p[SZ_UNTIL]
+
+    if not updates:
+        return
+
+    new_state = dataclasses.replace(target.dhw_state, **updates)
+    target.dhw_state = new_state
+
+    event = StateUpdatedEvent(
+        entity_id=target.id,
+        state=new_state,
+        correlation_id=getattr(msg, "correlation_id", uuid.uuid4()),
+        causation_id=getattr(msg, "message_id", uuid.uuid4()),
+    )
+    target.apply_state_update(event)
+
+
+def _update_temperature_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Translate temperature data into a frozen StateUpdatedEvent.
+
+    :param target: Target entity to update.
+    :type target: Any
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    temp_state = getattr(target, "temp_state", None)
+    if temp_state is None or not dataclasses.is_dataclass(temp_state):
+        return
+
+    updates: dict[str, Any] = {}
+
+    if SZ_TEMPERATURE in p:
+        target_id = getattr(target, "id", str(target))
+        src_id = getattr(msg.src, "id", str(msg.src))
+
+        # Legacy Parity: Physical sensors only track their own local sensor readings.
+        # We must ignore Zone temperature syncs sent TO them by the Controller.
+        if getattr(target, "_SLUG", "") in ("TRV", "THM") and src_id != target_id:
+            pass
+        else:
+            updates[SZ_TEMPERATURE] = p[SZ_TEMPERATURE]
+
+    if "setpoint" in p:
+        updates[SZ_SETPOINT] = p[SZ_SETPOINT]
+
+    if not updates:
+        return
+
+    new_state = dataclasses.replace(target.temp_state, **updates)
+    event = StateUpdatedEvent(
+        entity_id=getattr(target, "id", "unknown"),
+        state=new_state,
+        correlation_id=getattr(msg, "correlation_id", uuid.uuid4()),
+        causation_id=getattr(msg, "message_id", uuid.uuid4()),
+    )
+    target.apply_state_update(event)
+
+
+def _update_demand_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Translate demand data into a frozen StateUpdatedEvent.
+
+    :param target: Target entity to update.
+    :type target: Any
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    demand_state = getattr(target, "demand_state", None)
+    if demand_state is None or not dataclasses.is_dataclass(demand_state):
+        return
+
+    updates: dict[str, Any] = {}
+    if SZ_HEAT_DEMAND in p:
+        updates[SZ_HEAT_DEMAND] = p[SZ_HEAT_DEMAND]
+    if SZ_RELAY_DEMAND in p:
+        updates[SZ_RELAY_DEMAND] = p[SZ_RELAY_DEMAND]
+        updates["relay_active"] = float(p[SZ_RELAY_DEMAND]) > 0.0
+    if msg.code == Code._0009 and "failsafe_enabled" in p:
+        updates["relay_failsafe"] = p["failsafe_enabled"]
+
+    if not updates:
+        return
+
+    new_state = dataclasses.replace(target.demand_state, **updates)
+    event = StateUpdatedEvent(
+        entity_id=getattr(target, "id", "unknown"),
+        state=new_state,
+        correlation_id=getattr(msg, "correlation_id", uuid.uuid4()),
+        causation_id=getattr(msg, "message_id", uuid.uuid4()),
+    )
+    target.apply_state_update(event)
+
+
+def _update_faultlog_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Translate 0418 fault log data into a frozen StateUpdatedEvent.
+
+    This handles the immutable tuple appending tracking required by the
+    CQRS FaultLogState read-model container.
+
+    :param target: Target entity to update.
+    :type target: Any
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    if msg.code != Code._0418 or getattr(target, "state", None) is None:
+        return
+    if type(target.state).__name__ != "FaultLogState":
+        return
+
+    # Guard: Ensure the entry index exists in the parsed payload
+    if "log_idx" not in p:
+        return
+
+    try:
+        entry = FaultLogEntry.from_msg(msg)
+        current_entries = getattr(target.state, "entries", ())
+        # Append to the immutable tuple, safely removing stale matching timestamps
+        filtered = [e for e in current_entries if e.timestamp != entry.timestamp]
+        new_entries = tuple(filtered) + (entry,)
+
+        new_state = dataclasses.replace(target.state, entries=new_entries)
+
+        event = StateUpdatedEvent(
+            entity_id=getattr(target, "id", "unknown"),
+            state=new_state,
+            correlation_id=getattr(msg, "correlation_id", uuid.uuid4()),
+            causation_id=getattr(msg, "message_id", uuid.uuid4()),
+        )
+        target.apply_state_update(event)
+    except (AttributeError, KeyError, TypeError, ValueError) as err:
+        _LOGGER.warning("Failed to process fault log entry from msg %s: %s", msg, err)
+
+
+def _route_2411_to_fan(gwy: Gateway, msg: Message) -> None:
+    """Route a 2411 parameter message to its HvacVentilator aggregate root.
+
+    Phase 2.95 removed the ``HvacVentilator._handle_msg`` override that
+    previously invoked ``_handle_2411_message`` (which sets
+    ``_supports_2411`` and stores the parameter) and
+    ``_handle_initialized_callback`` (which fires the ramses_cc entity
+    creation callback).  Without this routing, FAN devices never advertise
+    2411 support, so ramses_cc never creates the ~15 parameter ``number``
+    entities (comfort temperature, etc.) — see ramses_cc issue 851.
+
+    This re-wires the 2411 handling into the CQRS ingestion pipeline (where
+    issue 639 wants domain logic to live) instead of restoring the leaky
+    ``_handle_msg`` override.  ``_handle_2411_message`` reads
+    ``msg.payload`` directly, so it is invoked once per FAN target, outside
+    the per-payload loop in ``process_state_updates``.
+
+    :param gwy: Gateway handling device entities.
+    :type gwy: Gateway
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    if getattr(msg, "verb", "") == "RQ":
+        return
+
+    registry = getattr(gwy, "device_registry", None)
+    if registry is None:
+        return
+
+    candidates: list[Any] = []
+    if msg.src is not None:
+        src_dev = registry.device_by_id.get(msg.src.id)
+        if src_dev is not None:
+            candidates.append(src_dev)
+    if msg.dst is not None:
+        dst_dev = registry.device_by_id.get(msg.dst.id)
+        if dst_dev is not None and dst_dev not in candidates:
+            candidates.append(dst_dev)
+
+    for dev in candidates:
+        if not isinstance(dev, HvacVentilator):
+            continue
+        try:
+            dev._handle_2411_message(msg)
+            dev._handle_initialized_callback()
+        except (exc.RamsesException, AttributeError, TypeError, ValueError) as err:
+            _LOGGER.error(
+                "Failed to route 2411 message to ventilator %s: %s",
+                dev.id,
+                err,
+            )
+
+
+def _update_schedule_state(target: Any, p: dict[str, Any], msg: Message) -> None:
+    """Route 0006 version and 0404 fragment packets to Schedule read-models.
+
+    :param target: Target entity.
+    :type target: Any
+    :param p: Parsed payload dictionary.
+    :type p: dict[str, Any]
+    :param msg: Message envelope.
+    :type msg: Message
+    """
+    if msg.code not in (Code._0006, Code._0404):
+        return
+
+    sched = getattr(target, "schedule", None)
+    if sched is not None and hasattr(sched, "process_schedule_msg"):
+        sched.process_schedule_msg(msg)
+
+
+async def process_state_updates(gwy: Gateway, msg: Message) -> None:
+    """Ingest message payloads into entity state read-models.
+
+    Acts as a Strangler Fig, intercepting decoded payloads and mapping
+    them directly into the new `StateUpdatedEvent` structures.
+
+    :param gwy: Gateway handling device registry and state.
+    :type gwy: Gateway
+    :param msg: Message envelope containing payload.
+    :type msg: Message
+    """
+    # Notify candidate devices of _last_msg_dtm and all binding devices of rcvd_msg
+    if registry := getattr(gwy, "device_registry", None):
+        for dev in list(registry.device_by_id.values()):
+            if dev.id in (getattr(msg.src, "id", None), getattr(msg.dst, "id", None)):
+                if hasattr(dev, "_last_msg_dtm"):
+                    dev._last_msg_dtm = msg.dtm
+            if (bm := getattr(dev, "_binding_manager", None)) and getattr(
+                bm, "is_binding", False
+            ):
+                bm.rcvd_msg(msg)
+
+    if not isinstance(msg.payload, (dict, list)):
+        return
+
+    # 2411 parameter messages are handled by the FAN aggregate root directly
+    # (they set _supports_2411 and fire the initialized callback).  This runs
+    # before the per-payload loop because _handle_2411_message reads
+    # msg.payload as a whole.  See ramses_cc issue 851.
+    if msg.code == Code._2411:
+        _route_2411_to_fan(gwy, msg)
+
+    payloads = msg.payload if isinstance(msg.payload, list) else [msg.payload]
+    with contextlib.suppress(exc.DeviceNotFoundError, exc.SchemaInconsistentError):
+        for p in payloads:
+            if isinstance(p, dict):
+                await update_topology_schema_state(gwy, p, msg)
+
+    # Legacy Parity: Request packets (RQ) do not contain state update telemetry.
+    if getattr(msg, "verb", "") == "RQ":
+        return
+
+    for p in payloads:
+        if not isinstance(p, dict):
+            continue
+        targets = _resolve_logical_targets(gwy, msg, p)
+        for target in targets:
+            _update_system_state(target, p, msg)
+            _update_hvac_state(target, p, msg)
+            _update_dhw_state(target, p, msg)
+            _update_temperature_state(target, p, msg)
+            _update_demand_state(target, p, msg)
+            _update_faultlog_state(target, p, msg)
+            _update_schedule_state(target, p, msg)
